requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
airtable-python-wrapper==0.13.0
annotated-types==0.7.0
anyascii==0.3.2
anyio==4.6.2.post1
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
coderedcms==4.0.1
decorator==5.1.1
defusedxml==0.7.1
Django==5.1.2
django-bootstrap5==24.2
django-debug-toolbar==4.4.6
django-eventtools==1.0.3
django-extensions==3.2.3
django-filter==24.3
django-modelcluster==6.3
django-permissionedforms==0.1
django-taggit==5.0.1
django-treebeard==4.7.1
djangorestframework==3.15.2
draftjs-exporter==5.0.0
et-xmlfile==1.1.0
exceptiongroup==1.2.2
filetype==1.2.0
frozenlist==1.4.1
future==1.0.0
geocoder==1.38.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
icalendar==5.0.13
idna==3.10
jsonpatch==1.33
jsonpointer==3.0.0
l18n==2021.3
laces==0.1.1
langchain==0.3.3
langchain-core==0.3.10
langchain-text-splitters==0.3.0
langsmith==0.1.135
multidict==6.1.0
numpy==1.26.4
openpyxl==3.1.5
orjson==3.10.7
packaging==24.1
pillow==10.4.0
pillow_heif==0.18.0
polib==1.2.0
propcache==0.2.0
pydantic==2.9.2
pydantic_core==2.23.4
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
ratelim==0.1.6
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.35
sqlparse==0.5.1
telepath==0.3.1
tenacity==8.5.0
tiktoken==0.8.0
typing_extensions==4.12.2
urllib3==2.2.3
wagtail==6.2.2
wagtail-ai==1.1.1
wagtail-airtable==0.7.0
wagtail-cache==2.5.1
wagtail-localize==1.10
wagtail-seo==2.5.0
wagtailmedia==0.15.2
Willow==1.8.0
yarl==1.15.2
